Refactored end to end tests -extracted common code to a base class
